Delete old db, add functions for storing server config
